Refactored commands and ChatBase for better organization
