+++ app/commands/ai_commands.py
@@ -1,6 +1,6 @@
 import datetime
 import os
-from telegram import Update, constants
+from telegram import Update
 from telegram.ext import ContextTypes, MessageHandler, CommandHandler, filters
 from app.app import AyanamiApp
 from models.command_base import CommandBase
@@ -75,19 +75,6 @@ class ResetCommand(CommandBase):
     def create(self):
         return CommandHandler(self.name, self.handle)
     
-class PingCommand(CommandBase):
-    def __init__(self, name, app: AyanamiApp):
-        super().__init__(name,)
-        self.app = app
-
-    async def handle(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
-        user = update.message.from_user
-        if user != None and self.app.is_authorized(user.id):
-            await self.chat.send_message(context=context, chat_id=update.effective_chat.id, text="pong")
-        
-    def create(self):
-        return CommandHandler(self.name, self.handle)
-
 class ChangeAICommand(CommandBase):
     def __init__(self, name, app: AyanamiApp, config):
         super().__init__(name,)
