+++ chat/telegramchat.py
@@ -13,3 +13,8 @@ class TelegramChat(ChatBase):
 
     def run(self):
         self.telegram_app.run_polling()
+
+    async def send_message(self, context, chat_id, text, connect_timeout=60):
+        await context.bot.send_message(chat_id=chat_id, 
+                                       text=text, 
+                                       connect_timeout=connect_timeout)
